pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.9.2"
description = "Educational project #2 on Hexlet platform (spec: Python developer)"
authors = ["Evgeny Sokolov <sokolovsolod@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"


[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
